pyproject.toml
[project]
name = "alteryx-server-py"
authors = [
    {name = "Paul Houghton", email = "paul.houghton@theinformationlab.co.uk"}
]
version = "0.2.0"
description = "Modern Python client for Alteryx Server API V3 with sync/async support"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["alteryx", "server", "api", "gallery", "workflow", "ci-cd", "automation"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking",
]

dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "hatch-vcs>=0.4.0",
    "hatch>=1.13.0",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
]

[project.urls]
"Documentation" = "https://github.com/Sivivatu/AlteryxGalleryAPI#readme"
"Repository" = "https://github.com/Sivivatu/AlteryxGalleryAPI"
"Bug Tracker" = "https://github.com/Sivivatu/AlteryxGalleryAPI/issues"

[project.scripts]

[build-system]
requires = ["uv_build>=0.9.22,<1"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 125
target-version = "py312"
force-exclude = true
extend-exclude = [
    "src/alteryx_gallery_api",
    "src/archive_AlteryxGallery",
    "scratch.py",
    "tests/archive_test_gallery_connection.py",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
